+++ async-http-client/src/main/scala/org/http4s/client/asynchttpclient/AsyncHttpClient.scala
@@ -38,7 +38,7 @@ object AsyncHttpClient {
     * @param ec The ExecutionContext to run responses on
     */
   def apply[F[_]](config: AsyncHttpClientConfig = defaultConfig)(
-      implicit F: Effect[F],
+      implicit F: ConcurrentEffect[F],
       ec: ExecutionContext): Client[F] = {
     val client = new DefaultAsyncHttpClient(config)
     Client(
@@ -61,18 +61,20 @@ object AsyncHttpClient {
     * shutdown when the stream terminates.
     */
   def stream[F[_]](config: AsyncHttpClientConfig = defaultConfig)(
-      implicit F: Effect[F],
+      implicit F: ConcurrentEffect[F],
       ec: ExecutionContext): Stream[F, Client[F]] =
-    Stream.bracket(F.delay(apply(config)))(c => Stream.emit(c), _.shutdown)
+    Stream.bracket(F.delay(apply(config)))(_.shutdown)
 
   private def asyncHandler[F[_]](
-      cb: Callback[DisposableResponse[F]])(implicit F: Effect[F], ec: ExecutionContext) =
+      cb: Callback[DisposableResponse[F]])(implicit F: ConcurrentEffect[F], ec: ExecutionContext) =
     new StreamedAsyncHandler[Unit] {
       var state: State = State.CONTINUE
       var dr: DisposableResponse[F] =
         DisposableResponse[F](Response(), F.delay(state = State.ABORT))
 
       override def onStream(publisher: Publisher[HttpResponseBodyPart]): State = {
+        implicit val timer: Timer[F] = Timer.derive[F](F, IO.timer(ec))
+
         // backpressure is handled by requests to the reactive streams subscription
         StreamSubscriber[F, HttpResponseBodyPart]
           .map { subscriber =>
@@ -116,8 +118,9 @@ object AsyncHttpClient {
       }
     }
 
-  private def toAsyncRequest[F[_]: Effect](request: Request[F])(
-      implicit ec: ExecutionContext): AsyncRequest = {
+  private def toAsyncRequest[F[_]: ConcurrentEffect](request: Request[F])(
+      implicit
+      ec: ExecutionContext): AsyncRequest = {
     val headers = new DefaultHttpHeaders
     for (h <- request.headers)
       headers.add(h.name.toString, h.value)
@@ -128,8 +131,11 @@ object AsyncHttpClient {
       .build()
   }
 
-  private def getBodyGenerator[F[_]: Effect](req: Request[F])(
-      implicit ec: ExecutionContext): BodyGenerator = {
+  private def getBodyGenerator[F[_]](req: Request[F])(
+      implicit
+      F: ConcurrentEffect[F],
+      ec: ExecutionContext): BodyGenerator = {
+    implicit val timer: Timer[F] = Timer.derive[F](F, IO.timer(ec))
     val publisher = StreamUnicastPublisher(
       req.body.chunks.map(chunk => Unpooled.wrappedBuffer(chunk.toArray)))
     if (req.isChunked) new ReactiveStreamsBodyGenerator(publisher, -1)