+++ async-http-client/src/test/scala/org/http4s/client/asynchttpclient/AsyncHttpClientSpec.scala
@@ -2,9 +2,10 @@ package org.http4s
 package client
 package asynchttpclient
 
-import cats.effect.{Effect, IO}
+import cats.effect.{ConcurrentEffect, IO}
+import scala.concurrent.ExecutionContext.Implicits.global
 
 class AsyncHttpClientSpec
     extends ClientRouteTestBattery(
       "AsyncHttpClient",
-      AsyncHttpClient()(Effect[IO], Http4sSpec.TestExecutionContext))
+      AsyncHttpClient()(ConcurrentEffect[IO], Http4sSpec.TestExecutionContext))