+++ blaze-server/src/main/scala/org/http4s/server/blaze/ProtocolSelector.scala
@@ -2,7 +2,7 @@ package org.http4s
 package server
 package blaze
 
-import cats.effect.Effect
+import cats.effect.ConcurrentEffect
 import java.nio.ByteBuffer
 import javax.net.ssl.SSLEngine
 import org.http4s.blaze.http.http2.{DefaultFlowStrategy, Http2Settings}
@@ -13,7 +13,7 @@ import scala.concurrent.duration.Duration
 
 /** Facilitates the use of ALPN when using blaze http2 support */
 private[blaze] object ProtocolSelector {
-  def apply[F[_]: Effect](
+  def apply[F[_]: ConcurrentEffect](
       engine: SSLEngine,
       routes: HttpRoutes[F],
       maxRequestLineLen: Int,