+++ client/src/main/scala/org/http4s/client/middleware/RequestLogger.scala
@@ -4,11 +4,11 @@ package middleware
 
 import cats.data.Kleisli
 import cats.effect._
+import cats.effect.concurrent.Ref
 import cats.implicits._
 import fs2._
 import org.http4s.util.CaseInsensitiveString
 import org.log4s._
-import scala.concurrent.ExecutionContext
 
 /**
   * Simple Middleware for Logging Requests As They Are Processed
@@ -16,18 +16,18 @@ import scala.concurrent.ExecutionContext
 object RequestLogger {
   private[this] val logger = getLogger
 
-  def apply[F[_]: Effect](
+  def apply[F[_]: Concurrent](
       logHeaders: Boolean,
       logBody: Boolean,
       redactHeadersWhen: CaseInsensitiveString => Boolean = Headers.SensitiveHeaders.contains
-  )(client: Client[F])(implicit ec: ExecutionContext = ExecutionContext.global): Client[F] =
-    client.copy(open = Kleisli { req =>
-      if (!logBody)
-        Logger.logMessage[F, Request[F]](req)(logHeaders, logBody)(logger.info(_)) *> client.open(
-          req)
-      else
-        async.refOf[F, Vector[Segment[Byte, Unit]]](Vector.empty[Segment[Byte, Unit]]).flatMap {
-          vec =>
+  )(client: Client[F]): Client[F] =
+    client.copy(open = Kleisli {
+      req =>
+        if (!logBody)
+          Logger.logMessage[F, Request[F]](req)(logHeaders, logBody)(logger.info(_)) *> client.open(
+            req)
+        else
+          Ref[F].of(Vector.empty[Segment[Byte, Unit]]).flatMap { vec =>
             val newBody = Stream
               .eval(vec.get)
               .flatMap(v => Stream.emits(v).covary[F])
@@ -36,7 +36,7 @@ object RequestLogger {
             val changedRequest = req.withBodyStream(
               req.body
               // Cannot Be Done Asynchronously - Otherwise All Chunks May Not Be Appended Previous to Finalization
-                .observe(_.segments.flatMap(s => Stream.eval_(vec.modify(_ :+ s))))
+                .observe(_.segments.flatMap(s => Stream.eval_(vec.update(_ :+ s))))
                 .onFinalize(
                   Logger.logMessage[F, Request[F]](req.withBodyStream(newBody))(
                     logHeaders,
@@ -46,6 +46,6 @@ object RequestLogger {
             )
 
             client.open(changedRequest)
-        }
+          }
     })
 }