+++ examples/jetty/src/main/scala/com/example/http4s/jetty/JettyExample.scala
@@ -13,17 +13,17 @@ import scala.concurrent.ExecutionContext.Implicits.global
 
 object JettyExample extends JettyExampleApp[IO]
 
-class JettyExampleApp[F[_]: Effect] extends StreamApp[F] with Http4sDsl[F] {
+class JettyExampleApp[F[_]: ConcurrentEffect] extends StreamApp[F] with Http4sDsl[F] {
   val metricsRegistry: MetricRegistry = new MetricRegistry
   val metrics: HttpMiddleware[F] = Metrics[F](metricsRegistry)
 
-  def stream(args: List[String], requestShutdown: F[Unit]): Stream[F, ExitCode] =
-    Scheduler(corePoolSize = 2).flatMap { implicit scheduler =>
-      JettyBuilder[F]
-        .bindHttp(8080)
-        .mountService(metrics(new ExampleService[F].service), "/http4s")
-        .mountService(metricsService(metricsRegistry), "/metrics")
-        .mountFilter(NoneShallPass, "/http4s/science/black-knight/*")
-        .serve
-    }
+  def stream(args: List[String], requestShutdown: F[Unit]): Stream[F, ExitCode] = {
+    implicit val timer = Timer.derive[F]
+    JettyBuilder[F]
+      .bindHttp(8080)
+      .mountService(metrics(new ExampleService[F].service), "/http4s")
+      .mountService(metricsService(metricsRegistry), "/metrics")
+      .mountFilter(NoneShallPass, "/http4s/science/black-knight/*")
+      .serve
+  }
 }