+++ examples/tomcat/src/main/scala/com/example/http4s/tomcat/TomcatExample.scala
@@ -12,17 +12,17 @@ import scala.concurrent.ExecutionContext.Implicits.global
 
 object TomcatExample extends TomcatExampleApp[IO]
 
-class TomcatExampleApp[F[_]: Effect] extends StreamApp[F] {
+class TomcatExampleApp[F[_]: ConcurrentEffect] extends StreamApp[F] {
   val metricsRegistry: MetricRegistry = new MetricRegistry
   val metrics: HttpMiddleware[F] = Metrics[F](metricsRegistry)
 
-  def stream(args: List[String], requestShutdown: F[Unit]): Stream[F, ExitCode] =
-    Scheduler(corePoolSize = 2).flatMap { implicit scheduler =>
-      TomcatBuilder[F]
-        .bindHttp(8080)
-        .mountService(metrics(new ExampleService[F].service), "/http4s")
-        .mountService(metricsService(metricsRegistry), "/metrics/*")
-        .mountFilter(NoneShallPass, "/http4s/science/black-knight/*")
-        .serve
-    }
+  def stream(args: List[String], requestShutdown: F[Unit]): Stream[F, ExitCode] = {
+    implicit val timer = Timer.derive[F]
+    TomcatBuilder[F]
+      .bindHttp(8080)
+      .mountService(metrics(new ExampleService[F].service), "/http4s")
+      .mountService(metricsService(metricsRegistry), "/metrics/*")
+      .mountFilter(NoneShallPass, "/http4s/science/black-knight/*")
+      .serve
+  }
 }