+++ server/src/main/scala/org/http4s/server/middleware/Logger.scala
@@ -7,7 +7,6 @@ import cats.effect._
 import fs2._
 import org.http4s.util.CaseInsensitiveString
 import org.log4s.getLogger
-import scala.concurrent.ExecutionContext
 
 /**
   * Simple Middleware for Logging All Requests and Responses
@@ -15,13 +14,13 @@ import scala.concurrent.ExecutionContext
 object Logger {
   private[this] val logger = getLogger
 
-  def apply[F[_]: Effect](
+  def apply[F[_]: Concurrent](
       logHeaders: Boolean,
       logBody: Boolean,
       redactHeadersWhen: CaseInsensitiveString => Boolean = Headers.SensitiveHeaders.contains,
       logAction: String => Unit = logger.info(_)
-  )(@deprecatedName('httpService) http: Kleisli[F, Request[F], Response[F]])(
-      implicit ec: ExecutionContext): Kleisli[F, Request[F], Response[F]] =
+  )(@deprecatedName('httpService) http: Kleisli[F, Request[F], Response[F]])
+    : Kleisli[F, Request[F], Response[F]] =
     ResponseLogger(logHeaders, logBody, redactHeadersWhen, logAction)(
       RequestLogger(logHeaders, logBody, redactHeadersWhen, logAction)(http)
     )